+++ R/export.R
@@ -123,7 +123,7 @@ export <- function(x, file, format, ...) {
     fmt <- get_type(fmt)
     outfile <- file
     if (fmt %in% c("gz", "gzip")) {
-        fmt <- file_ext(file_path_sans_ext(file, compression = FALSE))
+        fmt <- tools::file_ext(tools::file_path_sans_ext(file, compression = FALSE))
         file <- gzfile(file, "w")
         on.exit(close(file))
     }