+++ R/gather_attrs.R
@@ -11,6 +11,7 @@
 #' str(attributes(e))
 #' str(g)
 #' @seealso \code{\link{import}}, \code{\link{characterize}}
+#' @importFrom stats setNames
 #' @export
 gather_attrs <- function(x) {
     if (!inherits(x, "data.frame")) {
@@ -28,8 +29,8 @@ gather_attrs <- function(x) {
     }
     if (any(sapply(varattrs, length))) {
         attrnames <- sort(unique(unlist(lapply(varattrs, names))))
-        outattrs <- setNames(lapply(attrnames, function(z) {
-            setNames(lapply(varattrs, `[[`, z), names(x))
+        outattrs <- stats::setNames(lapply(attrnames, function(z) {
+            stats::setNames(lapply(varattrs, `[[`, z), names(x))
         }), attrnames)
         attributes(x) <- c(dfattrs, outattrs)
     }