+++ NEWS.md
@@ -1,3 +1,8 @@
+# rio 0.5.14
+
+ * Changes to `gather_attrs()` for haven 2.0.0 release.
+ * Some style guide changes to code.
+
 # rio 0.5.13
 
  * Allow `import()` of objects other than data frames from R-serialized (.rds and .rdata) files. Also, export of such objects to .rds files is supported, as previously intended. (#183, h/t Nicholas Jhirad)